SCMP dedicated DatagramChannel
