+++ src/main/java/org/scion/DatagramChannel.java
@@ -18,37 +18,26 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.net.*;
 import java.nio.ByteBuffer;
-import java.nio.channels.AlreadyConnectedException;
 import java.nio.channels.ByteChannel;
-import java.nio.channels.ClosedChannelException;
 import java.nio.channels.NotYetConnectedException;
-import java.time.Instant;
-import java.util.function.Consumer;
-import org.scion.internal.ExtensionHeader;
 import org.scion.internal.InternalConstants;
-import org.scion.internal.PathHeaderParser;
 import org.scion.internal.ScionHeaderParser;
-import org.scion.internal.ScmpParser;
 
-public class DatagramChannel implements ByteChannel, Closeable {
+public class DatagramChannel extends AbstractDatagramChannel<DatagramChannel>
+    implements ByteChannel, Closeable {
 
-  private final java.nio.channels.DatagramChannel channel;
-  private boolean isConnected = false;
-  private InetSocketAddress connection;
-  private RequestPath path;
-  private final ByteBuffer buffer;
-  private boolean cfgReportFailedValidation = false;
-  private PathPolicy pathPolicy = PathPolicy.DEFAULT;
-  private ScionService service;
-  private int cfgExpirationSafetyMargin =
-      ScionUtil.getPropertyOrEnv(
-          Constants.PROPERTY_PATH_EXPIRY_MARGIN,
-          Constants.ENV_PATH_EXPIRY_MARGIN,
-          Constants.DEFAULT_PATH_EXPIRY_MARGIN);
+  private final ByteBuffer bufferReceive;
+  private final ByteBuffer bufferSend;
 
-  private Consumer<Scmp.EchoResult> pingListener;
-  private Consumer<Scmp.TracerouteResult> traceListener;
-  private Consumer<Scmp.Message> errorListener;
+  protected DatagramChannel() throws IOException {
+    this(null);
+  }
+
+  protected DatagramChannel(ScionService service) throws IOException {
+    super(service);
+    this.bufferReceive = ByteBuffer.allocateDirect(getOption(StandardSocketOptions.SO_RCVBUF));
+    this.bufferSend = ByteBuffer.allocateDirect(getOption(StandardSocketOptions.SO_SNDBUF));
+  }
 
   public static DatagramChannel open() throws IOException {
     return new DatagramChannel();
@@ -58,272 +47,26 @@ public class DatagramChannel implements ByteChannel, Closeable {
     return new DatagramChannel(service);
   }
 
-  protected DatagramChannel() throws IOException {
-    this(null);
-  }
-
-  protected DatagramChannel(ScionService service) throws IOException {
-    this.channel = java.nio.channels.DatagramChannel.open();
-    this.service = service;
-    // TODO snd/rcv
-    int size =
-        Math.max(
-            channel.getOption(StandardSocketOptions.SO_RCVBUF),
-            channel.getOption(StandardSocketOptions.SO_SNDBUF));
-    this.buffer = ByteBuffer.allocateDirect(size);
-  }
-
-  /**
-   * Set the path policy. The default path policy is set in PathPolicy.DEFAULT, which currently
-   * means to use the first path returned by the daemon or control service. If the channel is
-   * connected, this method will request a new path using the new policy.
-   *
-   * @param pathPolicy the new path policy
-   * @see PathPolicy#DEFAULT
-   */
-  public synchronized void setPathPolicy(PathPolicy pathPolicy) throws IOException {
-    this.pathPolicy = pathPolicy;
-    if (path != null) {
-      updatePath(path);
-    }
-  }
-
-  public synchronized PathPolicy getPathPolicy() {
-    return this.pathPolicy;
-  }
-
-  public synchronized void setService(ScionService service) {
-    this.service = service;
-  }
-
-  public synchronized ScionService getService() {
-    if (service == null) {
-      service = Scion.defaultService();
-    }
-    return this.service;
-  }
-
-  public synchronized DatagramChannel bind(InetSocketAddress address) throws IOException {
-    channel.bind(address);
-    return this;
-  }
-
   // TODO we return `void` here. If we implement SelectableChannel
   //  this can be changed to return SelectableChannel.
+  @Override
   public synchronized void configureBlocking(boolean block) throws IOException {
-    channel.configureBlocking(block);
+    super.configureBlocking(block);
   }
 
+  @Override
   public synchronized boolean isBlocking() {
-    return channel.isBlocking();
-  }
-
-  public synchronized InetSocketAddress getLocalAddress() throws IOException {
-    return (InetSocketAddress) channel.getLocalAddress();
-  }
-
-  public synchronized void disconnect() throws IOException {
-    channel.disconnect();
-    connection = null;
-    isConnected = false;
-    path = null;
-  }
-
-  @Override
-  public synchronized boolean isOpen() {
-    return channel.isOpen();
-  }
-
-  @Override
-  // TODO not synchronized yet, think about a more fine grained lock (see JDK Channel impl)
-  public void close() throws IOException {
-    channel.disconnect();
-    channel.close();
-    isConnected = false;
-    connection = null;
-    path = null;
-  }
-
-  /**
-   * Connect to a destination host. Note: - A SCION channel will internally connect to the next
-   * border router (first hop) instead of the remote host.
-   *
-   * <p>NB: This method does internally no call {@link java.nio.channels.DatagramChannel}.connect(),
-   * instead it calls bind(). That means this method does NOT perform any additional security checks
-   * associated with connect(), only those associated with bind().
-   *
-   * @param addr Address of remote host.
-   * @return This channel.
-   * @throws IOException for example when the first hop (border router) cannot be connected.
-   */
-  public synchronized DatagramChannel connect(SocketAddress addr) throws IOException {
-    checkConnected(false);
-    if (!(addr instanceof InetSocketAddress)) {
-      throw new IllegalArgumentException(
-          "connect() requires an InetSocketAddress or a ScionSocketAddress.");
-    }
-    return connect(pathPolicy.filter(getService().getPaths((InetSocketAddress) addr)));
-  }
-
-  /**
-   * Connect to a destination host. Note: - A SCION channel will internally connect to the next
-   * border router (first hop) instead of the remote host. - The path will be replaced with a new
-   * path once it is expired.
-   *
-   * <p>NB: This method does internally no call {@link java.nio.channels.DatagramChannel}.connect(),
-   * instead it calls bind(). That means this method does NOT perform any additional security checks
-   * associated with connect(), only those associated with bind().
-   *
-   * @param path Path to the remote host.
-   * @return This channel.
-   * @throws IOException for example when the first hop (border router) cannot be connected.
-   */
-  public synchronized DatagramChannel connect(RequestPath path) throws IOException {
-    checkConnected(false);
-    this.path = path;
-    isConnected = true;
-    connection = path.getFirstHopAddress();
-    channel.connect(connection);
-    return this;
-  }
-
-  /**
-   * Get the currently connected path.
-   *
-   * @return the current Path or `null` if not path is connected.
-   */
-  public synchronized Path getCurrentPath() {
-    return path;
+    return super.isBlocking();
   }
 
   public synchronized ResponsePath receive(ByteBuffer userBuffer) throws IOException {
-    ResponsePath receivePath = receiveFromChannel(InternalConstants.HdrTypes.UDP);
+    ResponsePath receivePath = receiveFromChannel(bufferReceive, InternalConstants.HdrTypes.UDP);
     if (receivePath == null) {
       return null; // non-blocking, nothing available
     }
-    ScionHeaderParser.extractUserPayload(buffer, userBuffer);
-    buffer.clear();
+    ScionHeaderParser.extractUserPayload(bufferReceive, userBuffer);
+    bufferReceive.clear();
     return receivePath;
-  }
-
-  synchronized Scmp.Message receiveScmp() throws IOException {
-    ResponsePath receivePath = receiveFromChannel(InternalConstants.HdrTypes.SCMP);
-    if (receivePath == null) {
-      return null; // non-blocking, nothing available
-    }
-    return receiveScmp(receivePath);
-  }
-
-  synchronized Scmp.Message receiveScmp(Scmp.Message msg) throws IOException {
-    ResponsePath receivePath = receiveFromChannel(InternalConstants.HdrTypes.SCMP);
-    if (receivePath == null) {
-      return null; // non-blocking, nothing available
-    }
-    msg.setPath(receivePath);
-    return receiveScmpInternal(msg);
-  }
-
-  private ResponsePath receiveFromChannel(InternalConstants.HdrTypes expectedHdrType)
-      throws IOException {
-    while (true) {
-      buffer.clear();
-      InetSocketAddress srcAddress = (InetSocketAddress) channel.receive(buffer);
-      if (srcAddress == null) {
-        // this indicates nothing is available - non-blocking mode
-        return null;
-      }
-      buffer.flip();
-
-      String validationResult = ScionHeaderParser.validate(buffer.asReadOnlyBuffer());
-      if (validationResult != null) {
-        if (cfgReportFailedValidation) {
-          throw new ScionException(validationResult);
-        }
-        continue;
-      }
-
-      InternalConstants.HdrTypes hdrType = ScionHeaderParser.extractNextHeader(buffer);
-      if (expectedHdrType == hdrType && hdrType == InternalConstants.HdrTypes.UDP) {
-        return ScionHeaderParser.extractResponsePath(buffer, srcAddress);
-      }
-
-      // From here on we use linear reading using the buffer's position() mechanism
-      buffer.position(ScionHeaderParser.extractHeaderLength(buffer));
-      if (hdrType == InternalConstants.HdrTypes.END_TO_END
-          || hdrType == InternalConstants.HdrTypes.HOP_BY_HOP) {
-        ExtensionHeader extHdr = ExtensionHeader.consume(buffer);
-        // Currently we are not doing much here except hoping for an SCMP header
-        hdrType = extHdr.nextHdr();
-        if (hdrType != InternalConstants.HdrTypes.SCMP) {
-          throw new UnsupportedOperationException("Extension header not supported: " + hdrType);
-        }
-      }
-
-      if (hdrType == expectedHdrType) {
-        return ScionHeaderParser.extractResponsePath(buffer, srcAddress);
-      }
-      receiveScmp(path);
-    }
-  }
-
-  private Object receiveNonDataPacket(InternalConstants.HdrTypes hdrType, ResponsePath path)
-      throws ScionException {
-    switch (hdrType) {
-      case HOP_BY_HOP:
-      case END_TO_END:
-        return receiveExtension(path);
-        // break;
-      case SCMP:
-        return receiveScmp(path);
-        // break;
-      default:
-        if (cfgReportFailedValidation) {
-          throw new ScionException("Unknown nextHdr: " + hdrType);
-        }
-    }
-    return null;
-  }
-
-  private Object receiveExtension(ResponsePath path) throws ScionException {
-    ExtensionHeader extHdr = ExtensionHeader.consume(buffer);
-    // Currently we are not doing much here except hoping for an SCMP header
-    return receiveNonDataPacket(extHdr.nextHdr(), path);
-  }
-
-  private Scmp.Message receiveScmp(Path path) {
-    Scmp.Message scmpMsg = ScmpParser.consume(buffer, path);
-    if (scmpMsg instanceof Scmp.EchoResult) {
-      if (pingListener != null) {
-        pingListener.accept((Scmp.EchoResult) scmpMsg);
-      }
-    } else if (scmpMsg instanceof Scmp.TracerouteResult) {
-      if (traceListener != null) {
-        traceListener.accept((Scmp.TracerouteResult) scmpMsg);
-      }
-    } else {
-      if (errorListener != null) {
-        errorListener.accept(scmpMsg);
-      }
-    }
-    return scmpMsg;
-  }
-
-  private Scmp.Message receiveScmpInternal(Scmp.Message holder) {
-    Scmp.Message scmpMsg = ScmpParser.consume(buffer, holder);
-    if (scmpMsg instanceof Scmp.EchoResult) {
-      if (pingListener != null) {
-        pingListener.accept((Scmp.EchoResult) scmpMsg);
-      }
-    } else if (scmpMsg instanceof Scmp.TracerouteResult) {
-      if (traceListener != null) {
-        traceListener.accept((Scmp.TracerouteResult) scmpMsg);
-      }
-    } else {
-      if (errorListener != null) {
-        errorListener.accept(scmpMsg);
-      }
-    }
-    return scmpMsg;
   }
 
   /**
@@ -342,7 +85,7 @@ public class DatagramChannel implements ByteChannel, Closeable {
     if (!(destination instanceof InetSocketAddress)) {
       throw new IllegalArgumentException("Address must be of type InetSocketAddress.");
     }
-    send(srcBuffer, pathPolicy.filter(getService().getPaths((InetSocketAddress) destination)));
+    send(srcBuffer, getPathPolicy().filter(getService().getPaths((InetSocketAddress) destination)));
   }
 
   /**
@@ -359,73 +102,12 @@ public class DatagramChannel implements ByteChannel, Closeable {
    */
   public synchronized Path send(ByteBuffer srcBuffer, Path path) throws IOException {
     // + 8 for UDP overlay header length
-    Path actualPath = buildHeader(path, srcBuffer.remaining() + 8, InternalConstants.HdrTypes.UDP);
-    buffer.put(srcBuffer);
-    buffer.flip();
-    channel.send(buffer, actualPath.getFirstHopAddress());
+    Path actualPath =
+        buildHeader(bufferSend, path, srcBuffer.remaining() + 8, InternalConstants.HdrTypes.UDP);
+    bufferSend.put(srcBuffer);
+    bufferSend.flip();
+    sendRaw(bufferSend, actualPath.getFirstHopAddress());
     return actualPath;
-  }
-
-  public synchronized void sendEchoRequest(Path path, int sequenceNumber, ByteBuffer data)
-      throws IOException {
-    // EchoHeader = 8 + data
-    int len = 8 + data.remaining();
-    Path actualPath = buildHeader(path, len, InternalConstants.HdrTypes.SCMP);
-    ScmpParser.buildScmpPing(buffer, getLocalAddress().getPort(), sequenceNumber, data);
-    buffer.flip();
-    channel.send(buffer, actualPath.getFirstHopAddress());
-  }
-
-  public synchronized Scmp.EchoResult sendEchoRequest(Scmp.EchoResult request) throws IOException {
-    // EchoHeader = 8 + data
-    int len = 8 + request.getData().length;
-    Path actualPath = buildHeader(path, len, InternalConstants.HdrTypes.SCMP);
-    request.setPath(actualPath);
-    ScmpParser.buildScmpPing(
-        buffer, getLocalAddress().getPort(), request.getSequenceNumber(), request.getData());
-    buffer.flip();
-    channel.send(buffer, actualPath.getFirstHopAddress());
-    return request;
-  }
-
-  void sendTracerouteRequest(Path path, int interfaceNumber, PathHeaderParser.Node node)
-      throws IOException {
-    // TracerouteHeader = 24
-    int len = 24;
-    // TODO we are modifying the raw path here, this is bad! It breaks concurrent usage.
-    //   we should only modify the outgoing packet.
-    byte[] raw = path.getRawPath();
-    raw[node.posHopFlags] = node.hopFlags;
-    Path actualPath = buildHeader(path, len, InternalConstants.HdrTypes.SCMP);
-    ScmpParser.buildScmpTraceroute(buffer, getLocalAddress().getPort(), interfaceNumber);
-    buffer.flip();
-    channel.send(buffer, actualPath.getFirstHopAddress());
-    // Clean up!  // TODO this is really bad!
-    raw[node.posHopFlags] = 0;
-  }
-
-  void sendRaw(ByteBuffer buffer, InetSocketAddress address) throws IOException {
-    channel.send(buffer, address);
-  }
-
-  public synchronized Consumer<Scmp.EchoResult> setEchoListener(
-      Consumer<Scmp.EchoResult> listener) {
-    Consumer<Scmp.EchoResult> old = pingListener;
-    pingListener = listener;
-    return old;
-  }
-
-  public synchronized Consumer<Scmp.TracerouteResult> setTracerouteListener(
-      Consumer<Scmp.TracerouteResult> listener) {
-    Consumer<Scmp.TracerouteResult> old = traceListener;
-    traceListener = listener;
-    return old;
-  }
-
-  public synchronized Consumer<Scmp.Message> setScmpErrorListener(Consumer<Scmp.Message> listener) {
-    Consumer<Scmp.Message> old = errorListener;
-    errorListener = listener;
-    return old;
   }
 
   /**
@@ -468,133 +150,14 @@ public class DatagramChannel implements ByteChannel, Closeable {
 
     int len = src.remaining();
     // + 8 for UDP overlay header length
-    buildHeader(path, len + 8, InternalConstants.HdrTypes.UDP);
-    buffer.put(src);
-    buffer.flip();
+    buildHeader(bufferSend, getCurrentPath(), len + 8, InternalConstants.HdrTypes.UDP);
+    bufferSend.put(src);
+    bufferSend.flip();
 
-    int sent = channel.write(buffer);
-    if (sent < buffer.limit() || buffer.remaining() > 0) {
+    int sent = channel().write(bufferSend);
+    if (sent < bufferSend.limit() || bufferSend.remaining() > 0) {
       throw new ScionException("Failed to send all data.");
     }
-    return len - buffer.remaining();
-  }
-
-  private void checkOpen() throws ClosedChannelException {
-    if (!channel.isOpen()) {
-      throw new ClosedChannelException();
-    }
-  }
-
-  private void checkConnected(boolean requiredState) {
-    if (requiredState != isConnected) {
-      if (isConnected) {
-        throw new AlreadyConnectedException();
-      } else {
-        throw new NotYetConnectedException();
-      }
-    }
-  }
-
-  public synchronized boolean isConnected() {
-    return isConnected;
-  }
-
-  public synchronized <T> DatagramChannel setOption(SocketOption<T> option, T t)
-      throws IOException {
-    if (option instanceof ScionSocketOptions.SciSocketOption) {
-      if (ScionSocketOptions.SN_API_THROW_PARSER_FAILURE.equals(option)) {
-        cfgReportFailedValidation = (Boolean) t;
-      } else if (ScionSocketOptions.SN_PATH_EXPIRY_MARGIN.equals(option)) {
-        cfgExpirationSafetyMargin = (Integer) t;
-      } else if (StandardSocketOptions.SO_RCVBUF.equals(option)) {
-        // TODO resize buf
-        channel.setOption(option, t);
-      } else if (StandardSocketOptions.SO_SNDBUF.equals(option)) {
-        // TODO resize buf
-        channel.setOption(option, t);
-      } else {
-        throw new UnsupportedOperationException();
-      }
-    } else {
-      channel.setOption(option, t);
-    }
-    return this;
-  }
-
-  /**
-   * @param path path
-   * @param payloadLength payload length
-   * @return argument path or a new path if the argument path was expired
-   * @throws IOException in case of IOException.
-   */
-  Path buildHeader(Path path, int payloadLength, InternalConstants.HdrTypes hdrType)
-      throws IOException {
-    buffer.clear();
-    long srcIA;
-    byte[] srcAddress;
-    int srcPort;
-    if (path instanceof ResponsePath) {
-      // We could get source IA, address and port locally, but it seems cleaner
-      // to get these from the inverted header.
-      ResponsePath rPath = (ResponsePath) path;
-      srcIA = rPath.getSourceIsdAs();
-      srcAddress = rPath.getSourceAddress();
-      srcPort = rPath.getSourcePort();
-    } else {
-      // For sending request path we need to have a valid local external address.
-      // For a valid local external address we need to be connected.
-      if (!isConnected) {
-        isConnected = true;
-        connection = path.getFirstHopAddress();
-        channel.connect(connection);
-      }
-
-      // check path expiration
-      path = ensureUpToDate((RequestPath) path);
-      srcIA = getService().getLocalIsdAs();
-      // Get external host address. This must be done *after* refreshing the path!
-      InetSocketAddress srcSocketAddress = (InetSocketAddress) channel.getLocalAddress();
-      srcAddress = srcSocketAddress.getAddress().getAddress();
-      srcPort = srcSocketAddress.getPort();
-    }
-
-    long dstIA = path.getDestinationIsdAs();
-    byte[] dstAddress = path.getDestinationAddress();
-    int dstPort = path.getDestinationPort();
-
-    byte[] rawPath = path.getRawPath();
-    ScionHeaderParser.write(
-        buffer, payloadLength, rawPath.length, srcIA, srcAddress, dstIA, dstAddress, hdrType);
-    ScionHeaderParser.writePath(buffer, rawPath);
-
-    // TODO move this outside of this method
-    if (hdrType == InternalConstants.HdrTypes.UDP) {
-      ScionHeaderParser.writeUdpOverlayHeader(buffer, payloadLength, srcPort, dstPort);
-    }
-
-    return path;
-  }
-
-  private Path ensureUpToDate(RequestPath path) throws IOException {
-    if (Instant.now().getEpochSecond() + cfgExpirationSafetyMargin <= path.getExpiration()) {
-      return path;
-    }
-    return updatePath(path);
-  }
-
-  private Path updatePath(RequestPath path) throws IOException {
-    // expired, get new path
-    RequestPath newPath = pathPolicy.filter(getService().getPaths(path));
-
-    if (isConnected) { // equal to !isBound at this point
-      if (!newPath.getFirstHopAddress().equals(this.connection)) {
-        // TODO only reconnect if firstHop is on different interface....?!
-        channel.disconnect();
-        this.connection = newPath.getFirstHopAddress();
-        channel.connect(this.connection);
-      }
-      this.path = newPath;
-    }
-    return newPath;
+    return len - bufferSend.remaining();
   }
 }
